So I wasn't sure on how to get the stuff you uploaded so here is my attempt to share with you what I did.
I basically started from scracth, made a basic templete for all of our 3 templete pages. From here we can add and remove stuff for the changes for each templete page.

It has a nav bar but its not pretty ( no colors yet). I will work on that tomorrow.

It has two colums, and two sample posts.

Also has a footer which I will play a bit more with as well.